Staging

See merge request ascratech/accounts!2

+++ application/modules/ac_vouchers/models/Voucher_model.php
@@ -5,7 +5,7 @@ class Voucher_model extends BaseModel {
   protected $table_name = "ac_vouchers";
   function __construct($data=array()) {
     parent::__construct($data);
-    $this->load->model(array('masters/period_model','masters/setting_model','transactions/Receipt_not_sent_argold_model'));
+    $this->load->model(array('masters/period_model', 'masters/setting_model', 'transactions/Receipt_not_sent_argold_model', 'transactions/ledger_model'));
   }
 
   public function validation_rules($klass='') {
@@ -208,8 +208,8 @@ class Voucher_model extends BaseModel {
   }
 
   public function after_save($action) {
-    if ($action=="update")
-      $this->ledger_model->delete('', array('table_id'=>$id), true);
+    if ($action=="update" && isset($id))
+      $this->ledger_model->delete('', array('table_id' => $id), true);
 
     $ledger_data = $this->ledger_model->set_data($this->attributes, $this->router->class, $this->prefix, $this->table_name);
     $ledger_obj = new ledger_model($ledger_data);
